sea hp

+++ Assets/Scripts/farm_manager.cs
@@ -1003,7 +1003,7 @@ public class farm_manager : MonoBehaviour
         PlayerPrefs.SetInt("Haenyeo_moving_speed", Haenyeo.moving_speed);
         PlayerPrefs.SetInt("Haenyeo_day", Haenyeo.day);
         PlayerPrefs.SetInt("Haenyeo_level", Haenyeo.level);
-        PlayerPrefs.SetInt("Haenyeo_hp", Haenyeo.hp);
+        PlayerPrefs.SetFloat("Haenyeo_hp", Haenyeo.hp);
         PlayerPrefs.SetString("lasttime", System.DateTime.Now.ToString());
         PlayerPrefs.SetInt("Quest_number", quest.quest_number);         //퀘스트 번호 저장
         PlayerPrefs.SetInt("Quest_state", quest.quest_state);           //퀘스트 상태 저장
@@ -1110,7 +1110,7 @@ public class farm_manager : MonoBehaviour
             Haenyeo.moving_speed = PlayerPrefs.GetInt("Haenyeo_moving_speed", 7);
             Haenyeo.day = PlayerPrefs.GetInt("Haenyeo_day", 1);
             Haenyeo.level = PlayerPrefs.GetInt("Haenyeo_level", 2);
-            Haenyeo.hp = PlayerPrefs.GetInt("Haenyeo_hp", 100);
+            Haenyeo.hp = PlayerPrefs.GetFloat("Haenyeo_hp", 100);
             bgm_volume.value = PlayerPrefs.GetFloat("Bgm_volume", 1);
             effect_volume.value = PlayerPrefs.GetFloat("Effect_volume", 1);
             quest.quest_number = PlayerPrefs.GetInt("Quest_number", 0);
@@ -1136,7 +1136,7 @@ public class farm_manager : MonoBehaviour
             Haenyeo.moving_speed = PlayerPrefs.GetInt("Haenyeo_moving_speed", 7);
             Haenyeo.day = PlayerPrefs.GetInt("Haenyeo_day", 1);
             Haenyeo.level = PlayerPrefs.GetInt("Haenyeo_level", 3); // 다해 : 바다 다 열려고 레벨 3으로 설정 해놨음
-            Haenyeo.hp = PlayerPrefs.GetInt("Haenyeo_hp", 100);
+            Haenyeo.hp = PlayerPrefs.GetFloat("Haenyeo_hp", 100);
             bgm_volume.value = PlayerPrefs.GetFloat("Bgm_volume", 1);
             effect_volume.value = PlayerPrefs.GetFloat("Effect_volume", 1);
             quest.quest_number = PlayerPrefs.GetInt("Quest_number", 0);
